Checkout: convert lib/checkout/validation to TypeScript and remove dead validation API
Convert client/lib/checkout/validation.js to TypeScript and remove the
unused rule-based payment validation API, keeping only getCreditCardType.

The rule-based API (validatePaymentDetails, paymentFieldRules,
getCreditCardFieldRules, getStripeElementsRules, tokenFieldRules,
mergeValidationRules) had no callers — field validation is now handled
by Stripe Elements / EBANX VGS. It was only referenced by its own test.

- Rename validation.js -> validation.ts.
- Keep getCreditCardType, still used to show the card-brand icon and
  drive number masking on the EBANX raw-input path.
- Remove the dead rule engine, its lodash usage (capitalize, isEmpty,
  mergeWith), and the now-unused @automattic/wpcom-checkout imports.
- Add a documented ambient `creditcards` module declaration scoped to
  the `card.type` surface we use (the package ships no types).
- Trim the barrel export, README, and unit test accordingly.
