+++ client/lib/checkout/validation.ts
@@ -1,0 +1,33 @@
+import creditcards from 'creditcards';
+
+/**
+ * Retrieves the type of credit card from the specified number.
+ * @param number - credit card number
+ * @returns the type of the credit card
+ * @see {@link http://en.wikipedia.org/wiki/Bank_card_number} for more information
+ */
+export function getCreditCardType( number?: string ): string | null {
+	if ( number ) {
+		number = number.replace( / /g, '' );
+
+		let cardType = creditcards.card.type( number, true );
+
+		if ( typeof cardType === 'undefined' ) {
+			return null;
+		}
+
+		// We already use 'amex' for American Express everywhere else
+		if ( cardType === 'American Express' ) {
+			cardType = 'amex';
+		}
+
+		// Normalize Diners as well
+		if ( cardType === 'Diners Club' ) {
+			cardType = 'diners';
+		}
+
+		return cardType.toLowerCase();
+	}
+
+	return null;
+}
